+++ tests/unit/no-top-level-side-effects.test.ts
@@ -460,6 +460,10 @@ const valid: RuleTester.ValidTestCase[] = [
       options: [options.commonjs]
     },
     {
+      code: `module.exports[foobar] = {};`,
+      options: [{...options.allowDerived, ...options.commonjs}]
+    },
+    {
       code: `module.exports.foo = bar();`,
       options: [{...options.commonjs, allowedCalls: ['bar']}]
     },
@@ -474,6 +478,10 @@ const valid: RuleTester.ValidTestCase[] = [
     {
       code: `exports.foobar = {};`,
       options: [options.commonjs]
+    },
+    {
+      code: `exports[foobar] = {};`,
+      options: [{...options.allowDerived, ...options.commonjs}]
     },
     {
       code: `exports.foo = bar();`,
@@ -512,11 +520,21 @@ const valid: RuleTester.ValidTestCase[] = [
       languageOptions: languageOptions.sourceTypeScript
     },
     {
+      code: `module.exports[foobar] = {};`,
+      options: [options.allowDerived],
+      languageOptions: languageOptions.sourceTypeScript
+    },
+    {
       code: `exports = {};`,
       languageOptions: languageOptions.sourceTypeScript
     },
     {
       code: `exports.foobar = {};`,
+      languageOptions: languageOptions.sourceTypeScript
+    },
+    {
+      code: `exports[foobar] = {};`,
+      options: [options.allowDerived],
       languageOptions: languageOptions.sourceTypeScript
     }
   ],
@@ -554,6 +572,11 @@ const valid: RuleTester.ValidTestCase[] = [
       languageOptions: languageOptions.sourceTypeModule
     },
     {
+      code: `module.exports[foobar] = {};`,
+      options: [{...options.allowDerived, ...options.commonjs}],
+      languageOptions: languageOptions.sourceTypeModule
+    },
+    {
       code: `exports = {};`,
       options: [options.commonjs],
       languageOptions: languageOptions.sourceTypeModule
@@ -562,6 +585,11 @@ const valid: RuleTester.ValidTestCase[] = [
       code: `exports.foobar = {};`,
       options: [options.commonjs],
       languageOptions: languageOptions.sourceTypeModule
+    },
+    {
+      code: `exports[foobar] = {};`,
+      options: [{...options.allowDerived, ...options.commonjs}],
+      languageOptions: languageOptions.sourceTypeScript
     }
   ],
 
@@ -1738,7 +1766,7 @@ const invalid: RuleTester.InvalidTestCase[] = [
     },
     {
       code: `console[log]('hello world');`,
-      options: [{allowedCalls: ['console.log']}],
+      options: [{allowedCalls: ['console.log'], ...options.allowDerived}],
       errors: [
         {
           messageId: '0',
@@ -2714,6 +2742,32 @@ const invalid: RuleTester.InvalidTestCase[] = [
           column: 1,
           endLine: 1,
           endColumn: 31
+        }
+      ]
+    },
+    {
+      code: `exports[foobar] = {};`,
+      languageOptions: languageOptions.sourceTypeScript,
+      errors: [
+        {
+          messageId: '0',
+          line: 1,
+          column: 1,
+          endLine: 1,
+          endColumn: 16
+        }
+      ]
+    },
+    {
+      code: `module.exports[foobar] = {};`,
+      languageOptions: languageOptions.sourceTypeScript,
+      errors: [
+        {
+          messageId: '0',
+          line: 1,
+          column: 1,
+          endLine: 1,
+          endColumn: 23
         }
       ]
     }
@@ -3934,7 +3988,7 @@ const invalid: RuleTester.InvalidTestCase[] = [
           line: 2,
           column: 9,
           endLine: 2,
-          endColumn: 41
+          endColumn: 32
         }
       ]
     },